Leaders in Array
You are given an integer array arr of size n. An element is considered a leader if it is greater than all the elements to its right. Your task is to find all such leader elements in the array.

Input:
An integer array arr of size n.
Example : 
arr = [16, 17, 4, 3, 5, 2]

Output:
Return an array containing all the leader elements in the order in which they appear in the original array.
Example:
Leaders: [17, 5, 2]

Explanation
16 is not a leader because 17 (to its right) is greater.
17 is a leader because there is no element to its right that is greater.
4 is not a leader because 5 (to its right) is greater.
3 is not a leader because 5 (to its right) is greater.
5 is a leader because there is no element to its right that is greater.
2 is a leader because there is no element to its right.

Constraints:
1 ≤ n ≤ 10^6
1 ≤ arr[i] ≤ 10^9

Edge Cases:
The array contains a single element, which is always a leader.
The array is sorted in descending order, making every element a leader.
The array is sorted in ascending order, making only the last element a leader.